move GetJob to package kubeclient and modify the conditions for job judgment
