+++ pkg/utils/dataload_test.go
@@ -21,7 +21,6 @@ import (
 
 	datav1alpha1 "github.com/fluid-cloudnative/fluid/api/v1alpha1"
 	"github.com/fluid-cloudnative/fluid/pkg/utils/fake"
-	batchv1 "k8s.io/api/batch/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
@@ -74,52 +73,6 @@ func TestGetDataLoad(t *testing.T) {
 		}
 	}
 
-}
-
-func TestGetDataLoadJob(t *testing.T) {
-	mockJobName := "fluid-test-job"
-	mockJobNamespace := "default"
-	initJob := &batchv1.Job{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      mockJobName,
-			Namespace: mockJobNamespace,
-		},
-	}
-
-	fakeClient := fake.NewFakeClient(initJob)
-
-	testCases := map[string]struct {
-		name      string
-		namespace string
-		wantName  string
-		notFound  bool
-	}{
-		"test get DataLoad Job case 1": {
-			name:      mockJobName,
-			namespace: mockJobNamespace,
-			wantName:  mockJobName,
-			notFound:  false,
-		},
-		"test get DataLoad Job case 2": {
-			name:      mockJobName + "not-exist",
-			namespace: mockJobNamespace,
-			wantName:  "",
-			notFound:  true,
-		},
-	}
-
-	for k, item := range testCases {
-		gotJob, err := GetJob(fakeClient, item.name, item.namespace)
-		if item.notFound {
-			if err == nil && gotJob != nil {
-				t.Errorf("%s check failure, want get err, but get nil", k)
-			}
-		} else {
-			if gotJob.Name != item.wantName {
-				t.Errorf("%s check failure, want DataLoad Job name:%s, got DataLoad Job name:%s", k, item.wantName, gotJob.Name)
-			}
-		}
-	}
 }
 
 func TestGetDataLoadReleaseName(t *testing.T) {
